move SurfaceRoughness to DataSurfaces.hh to avoid circular dependency

+++ src/EnergyPlus/SQLiteProcedures.cc
@@ -2208,7 +2208,7 @@ bool SQLite::Material::insertIntoSQLite(sqlite3_stmt *insertStmt)
     sqliteBindInteger(insertStmt, 1, number);
     sqliteBindText(insertStmt, 2, name);
     sqliteBindInteger(insertStmt, 3, static_cast<int>(group));
-    sqliteBindInteger(insertStmt, 4, roughness);
+    sqliteBindInteger(insertStmt, 4, static_cast<int>(roughness));
     sqliteBindDouble(insertStmt, 5, conductivity);
     sqliteBindDouble(insertStmt, 6, density);
     sqliteBindDouble(insertStmt, 7, isoMoistCap);
@@ -2238,7 +2238,7 @@ bool SQLite::Construction::insertIntoSQLite(sqlite3_stmt *insertStmt)
     sqliteBindDouble(insertStmt, 9, outsideAbsorpSolar);
     sqliteBindDouble(insertStmt, 10, insideAbsorpThermal);
     sqliteBindDouble(insertStmt, 11, outsideAbsorpThermal);
-    sqliteBindInteger(insertStmt, 12, outsideRoughness);
+    sqliteBindInteger(insertStmt, 12, static_cast<int>(outsideRoughness));
     sqliteBindLogical(insertStmt, 13, typeIsWindow);
     sqliteBindDouble(insertStmt, 14, uValue);
